refactoring ProgressChartWidget
